Feature added
Created code to target each 'Add to Cart' product buttons.
Created an ajax call to send product id to the server as data.
