+++ assets/scripts/browse/events.js
@@ -2,13 +2,14 @@
 const api = require('./api.js')
 const ui = require('./ui.js')
 
-const onGetAllProducts = event => {
-  event.preventDefault()
-  console.log('onGetAllProducts', event.target)
-  api.getAllProducts()
-    .then(ui.getAllProductsSuccess)
-    .catch(ui.getAllProductsFailure)
-}
+// Only needed to test front end/back end connection.
+// const onGetAllProducts = event => {
+//   event.preventDefault()
+//   console.log('onGetAllProducts', event.target)
+//   api.getAllProducts()
+//     .then(ui.getAllProductsSuccess)
+//     .catch(ui.getAllProductsFailure)
+// }
 
 const onGetAProduct = event => {
   event.preventDefault()
@@ -20,16 +21,31 @@ const onGetAProduct = event => {
 
 const onAddToCart = event => {
   event.preventDefault()
-  console.log(event.target)
-  api.addToCart()
+  console.log('inside onAddToCart')
+  // console.log(event.target.id)
+  console.log($(event.target).attr('id'))
+  const data = $(event.target).attr('id')
+  api.addToCart(data)
     .then(ui.addToCartSuccess)
     .catch(ui.addToCartFailure)
+}
+
+const onCheckout = event => {
+  event.preventDefault()
+  console.log(event.target)
+  console.log('onCheckout')
+  api.checkout()
+    .then(ui.checkoutSuccess)
+    .catch(ui.checkoutFailure)
 }
 
 const onGetWomensClothing = event => {
   event.preventDefault()
   api.womensClothingApi()
     .then(ui.getWomensClothingSuccess)
+    // .then($('.add-to-cart').on('click', onAddToCart))
+    // .then($('#productList').on('click', '.add-to-cart', onAddToCart))
+    // .then($('#productList').on('click', '.checkout', onCheckout))
 }
 
 const onGetMensClothing = event => {
@@ -37,18 +53,11 @@ const onGetMensClothing = event => {
   api.mensClothingApi()
     .then(ui.getMensClothingSuccess)
 }
-//   event.preventDefault()
-//   console.log(event.target)
-//   api.checkout()
-//     .then(ui.checkoutSuccess)
-//     .catch(ui.checkoutFailure)
-// }
 
 module.exports = {
-  onGetAllProducts,
   onGetAProduct,
   onAddToCart,
   onGetWomensClothing,
-  onGetMensClothing
-  // onCheckout
+  onGetMensClothing,
+  onCheckout
 }
