+++ assets/scripts/browse/ui.js
@@ -1,4 +1,5 @@
 'use strict'
+const store = require('./../store.js')
 
 const getAllProductsSuccess = data => {
   console.log('getAllProductsSuccess running', data)
@@ -30,26 +31,30 @@ const getWomensClothingSuccess = function (data) {
   const dataArray = data.products
   dataArray.forEach(function (product) {
     const imgurUrl = product.image
+    // const addToCart = $('.')
     content += `<div class="col-sm-4">
-      <div class="card">
-        <div class="card-body">
+      <div class="card" id="imageDiv">
+        <div class="card-body" id="imageDivalso">
           <h5 class="card-title"></h5>
           <img class="card-img" src='${imgurUrl}.jpg' alt="Card image">
           <p class="card-text"></p>
-          <button type='submit' class="btn btn-primary add-to-cart">Add to cart</button>
+          <button class="btn btn-primary add-to-cart" id="${product._id}">Add to cart</button>
           <a href="#" class="btn btn-primary" id="checkout">Checkout Now</a>
         </div>
       </div>
     </div>`
-    console.log(product._id)
-    console.log(product.image)
+    // store.product._id = product
+    // console.log(product)
+    // console.log(product.image)
   })
+  content += '</div>'
   $('#productList').html(content)
 }
 
 const getMensClothingSuccess = function (data) {
   $('#productList').html('')
   let content = '<div class="row">'
+  store.product = data.products
   const dataArray = data.products
   dataArray.forEach(function (product) {
     const imgurUrl = product.image
@@ -60,15 +65,31 @@ const getMensClothingSuccess = function (data) {
           <img class="card-img" src='${imgurUrl}.jpg' alt="Card image">
           <p class="card-text"></p>
           <button type='submit' class="btn btn-primary add-to-cart">Add to cart</button>
-          <a href="#" class="btn btn-primary" id="checkout">Checkout Now</a>
+          <button type='submit' class="btn btn-primary checkout">Checkout Now</button>
         </div>
       </div>
     </div>`
-    console.log(product._id)
-    console.log(product.image)
+    // console.log(product._id)
+    // console.log(product.image)
   })
   $('#productList').html(content)
 }
+
+const viewCartSuccess = data => {
+  $('#message').text('view cart successfully')
+  $('#message').removeClass()
+  $('#message').addClass('success')
+  $('#change-password').trigger('reset')
+  console.log('viewCartSuccess ran. Data is :', data)
+}
+
+const viewCartFailure = error => {
+  $('#message').text('Error on view cart')
+  $('#message').removeClass()
+  $('#message').addClass('failure')
+  console.error('viewCartFailure ran. Error is :', error)
+}
+
 module.exports = {
   getAllProductsSuccess,
   getAllProductsFailure,
@@ -77,5 +98,7 @@ module.exports = {
   addToCartSuccess,
   addToCartFailure,
   getWomensClothingSuccess,
-  getMensClothingSuccess
+  getMensClothingSuccess,
+  viewCartSuccess,
+  viewCartFailure
 }
