+++ assets/scripts/event.js
@@ -36,23 +36,23 @@ const onSignOut = event => {
     .catch(ui.signOutFailure)
 }
 
-const onAddToCart = event => {
-  event.preventDefault()
-  const data = getFormFields(event.target)
-  console.log(data)
-  api.addToCart(data)
-    .then(ui.addToCartSuccess)
-    .catch(ui.addToCartFailure)
-}
-
-const onViewCart = event => {
-  event.preventDefault()
-  const data = getFormFields(event.target)
-  console.log(data)
-  api.viewCart(data)
-    .then(ui.viewCartSuccess)
-    .catch(ui.viewCartFailure)
-}
+// const onAddToCart = event => {
+//   event.preventDefault()
+//   const data = getFormFields(event.target)
+//   console.log(data)
+//   api.addToCart(data)
+//     .then(ui.addToCartSuccess)
+//     .catch(ui.addToCartFailure)
+// }
+//
+// const onViewCart = event => {
+//   event.preventDefault()
+//   const data = getFormFields(event.target)
+//   console.log(data)
+//   api.viewCart(data)
+//     .then(ui.viewCartSuccess)
+//     .catch(ui.viewCartFailure)
+// }
 
 const onMbrSign = event => {
   event.preventDefault()
@@ -66,8 +66,8 @@ module.exports = {
   onSignIn,
   onChangePassword,
   onSignOut,
-  onAddToCart,
-  onViewCart,
+  // onAddToCart,
+  // onViewCart,
   onMbrSign
 
 }
